+++ source/platform/renderers/Software/Software.cpp
@@ -89,18 +89,19 @@ static constexpr int c_texsz = 256;   // TODO - fix this hack.
 
 Font::Status SoftwareResourceLoader::LoadFont(Font& font)
 {
-    const String fontFile = m_paths.GetPath(ResourcePaths::Type::Font, font.facename);
+    const String filename = m_paths.GetPath(ResourcePaths::Type::Font, font.facename);
     
-    FILE* f = fopen(fontFile.c_str(), "rb");
+    FILE* f = fopen(filename.c_str(), "rb");
     if (!f)
     {
+        Gwk::Log::Write(Log::Level::Error, "Font file not found: %s", filename.c_str());
         font.data = nullptr;
         font.status = Font::Status::ErrorFileNotFound;
         return font.status;
     }
     
     struct stat finfo;
-    const int rc = stat(fontFile.c_str(), &finfo);
+    const int rc = stat(filename.c_str(), &finfo);
     const size_t fsz = rc == 0 ? finfo.st_size : -1;
     assert(fsz > 0);
     
@@ -148,6 +149,7 @@ Texture::Status SoftwareResourceLoader::LoadTexture(Texture& texture)
     
     if (!data)
     {
+        Gwk::Log::Write(Log::Level::Error, "Texture file not found: %s", filename.c_str());                        
         texture.status = Texture::Status::ErrorFileNotFound;
         return texture.status;
     }